Moved functionality from view to model
